src/main/java/com/github/brunomndantas/jscrapper/support/parser/array/text/reference/ArrayReferenceCharacterTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.array.text.reference;

import com.github.brunomndantas.jscrapper.support.parser.array.text.ArrayTextParser;
import org.openqa.selenium.WebElement;

public class ArrayReferenceCharacterTextParser extends ArrayTextParser {

    public ArrayReferenceCharacterTextParser() {
        super(Character.class, null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return element.getText().charAt(0);
    }

}